add (check storage too low): added option to check if the storage is too low for required resources (so resources wont get sent/used by hero). auto-builds storage if it's too low.

+++ TbsCore/Helpers/IoHelperCore.cs
@@ -118,6 +118,8 @@ namespace TravBotSharp.Files.Helpers
                 .Where(x => x.Replace(CachePath + "\\", "").StartsWith(userFolder))
                 .ToArray();
 
+            if (removeFolders == null) return;
+
             for (int i = 0; i < removeFolders.Count(); i++)
             {
                 Directory.Delete(removeFolders[i], true);
